// components/runtime/src/ts/generated/internal.ts
export type IMessageSnapshot = {
  message: IMessageSnapshotFields;
  guildId?: string;
};

export type IMessageSnapshotFields = {
  attachments: Array<Attachment>;
  components: Array<IComponent>;
  content: string;
  editedTimestamp: number | null;
  embeds: Array<Embed>;
  flags?: MessageFlags;
  kind: MessageType;
  mentions: Array<IUserMention>;
  mentionRoles: Array<string>;
  timestamp: number;
};
// ...
export type IUser = {
  avatar: string | null;
  bot: boolean;
  discriminator: string;
  id: string;
  locale: string | null;
  username: string;
  globalName: string | null;
  premiumType: PremiumType | null;
  publicFlags: IUserFlags | null;
  system: boolean | null;
  banner: string | null;
  accentColor: number | null;
  primaryGuild: IPrimaryGuild | null;
};
// ...
export type IUserMention = { user: IUser; member: PartialMember | null };